+++ client/package.json
@@ -15,6 +15,7 @@
     "chart.js": "^2.7.1",
     "chart.piecelabel.js": "^0.8.2",
     "element-ui": "^2.0.2",
+    "google-charts": "^1.0.0",
     "jsonexport": "^2.0.11",
     "lodash": "^4.17.4",
     "tiny-cookie": "^2.0.2",
