+++ client/src/charts/BarChart.js
@@ -8,7 +8,7 @@ export default {
   extends: Bar,
   props: ['options'],
   mixins: [mixins.reactiveProp],
-  mounted() {
+  mounted () {
     this.renderChart(this.chartData, this.options)
   }
 }
